requirements.txt
# Core PyTorch
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Configuration
PyYAML>=6.0

# Utilities
tqdm>=4.60.0
psutil>=5.8.0

# Optional: Jupyter support
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: Advanced visualization
plotly>=5.0.0

# Financial data
yfinance>=0.2.0

# Optional: Experiment tracking
# wandb>=0.12.0
# mlflow>=1.20.0

# Development
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0